fixed bufferSize, removed useless makefile entries

+++ lexer.h
@@ -5,4 +5,4 @@ int setupLexer(FILE* fp);
 TOKEN* getNextToken();
 void cleanLexer();
 void initbuf(void);
-void setBufferSize(int newBufSize);
+void setBufferSize(long long newBufSize);
